Add definition period test to variables
